Move `NodeError` type into `error` module
