+++ src/lib.rs
@@ -39,7 +39,7 @@ use core::{
     slice::Iter,
 };
 
-use failure::{bail, Fail, Fallible};
+use failure::{bail, Fallible};
 
 #[cfg(feature = "par_iter")]
 use rayon::prelude::*;
@@ -53,6 +53,10 @@ use std::{
     slice::Iter,
 };
 
+pub use crate::error::NodeError;
+
+mod error;
+
 #[derive(PartialEq, Eq, PartialOrd, Ord, Copy, Clone, Debug, Hash)]
 #[cfg_attr(feature = "deser", derive(Deserialize, Serialize))]
 /// A node identifier within a particular `Arena`
@@ -65,55 +69,6 @@ impl fmt::Display for NodeId {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         write!(f, "{}", self.index1)
     }
-}
-
-#[derive(Debug, Fail)]
-/// Possible node failures
-pub enum NodeError {
-    #[fail(display = "Can not append a node to itself")]
-    AppendSelf,
-
-    #[fail(display = "Can not prepend a node to itself")]
-    PrependSelf,
-
-    #[fail(display = "Can not insert a node before itself")]
-    InsertBeforeSelf,
-
-    #[fail(display = "Can not insert a node after itself")]
-    InsertAfterSelf,
-
-    #[fail(display = "First child is already set")]
-    FirstChildAlreadySet,
-
-    #[fail(display = "Previous sibling is already set")]
-    PreviousSiblingAlreadySet,
-
-    #[fail(display = "Next sibling is already set")]
-    NextSiblingAlreadySet,
-
-    #[fail(display = "Previous sibling not equal current node")]
-    PreviousSiblingNotSelf,
-
-    #[fail(display = "Next sibling not equal current node")]
-    NextSiblingNotSelf,
-
-    #[fail(display = "First child not equal current node")]
-    FirstChildNotSelf,
-
-    #[fail(display = "Last child not equal current node")]
-    LastChildNotSelf,
-
-    #[fail(display = "Previous sibling is not set")]
-    PreviousSiblingNotSet,
-
-    #[fail(display = "Next sibling is not set")]
-    NextSiblingNotSet,
-
-    #[fail(display = "First child is not set")]
-    FirstChildNotSet,
-
-    #[fail(display = "Last child is not set")]
-    LastChildNotSet,
 }
 
 #[derive(PartialEq, Clone, Debug)]
